+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # TBD
 * Declare timeouts for tests
+* Make initializer & controller log levels independently configurable
 
 # 0.2.1
 * Fix tee suppressing exit code of the Docker image